refactor: 게임 결과 소프트코딩 적용 및 파일 네이밍 변경(gameConfig => baseballGame)
